Pkcs11Interop_ver3

+++ src/Pkcs11Interop/Pkcs11Interop/HighLevelAPI40/MechanismParams/CkGOST3410DeriveParams.cs
@@ -60,14 +60,14 @@ namespace Net.Pkcs11Interop.HighLevelAPI40.MechanismParams
             {
                 _lowLevelStruct.pPublicData = UnmanagedMemory.Allocate(publicData.Length);
                 UnmanagedMemory.Write(_lowLevelStruct.pPublicData, publicData);
-                _lowLevelStruct.ulPublicDataLen = Convert.ToUInt64(publicData.Length);
+                _lowLevelStruct.ulPublicDataLen = Convert.ToUInt32(publicData.Length);
             }
 
             if (UKM != null)
             {
                 _lowLevelStruct.pUKM = UnmanagedMemory.Allocate(UKM.Length);
                 UnmanagedMemory.Write(_lowLevelStruct.pUKM, UKM);
-                _lowLevelStruct.ulUkmLen = Convert.ToUInt64(UKM.Length);
+                _lowLevelStruct.ulUkmLen = Convert.ToUInt32(UKM.Length);
             }
 
         }
